app/requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==42.0.5
Django==5.0.4
djangorestframework==3.15.1
pillow==10.3.0
pycparser==2.22
PyMySQL==1.1.0
sqlparse==0.4.4
typing_extensions==4.10.0
mysqlclient==2.1.1
